# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for dynamic user segmentation for testing new functionality",
        "title": "user-segmentation api",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
        "/report": {
            "post": {
                "description": "Generate a CSV report for a specific month and year",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Generate CSV report",
                "operationId": "generateCsvReport",
                "parameters": [
                    {
                        "description": "CSV report request",
                        "name": "csvReport",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.CsvReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated CSV report",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.CsvReportResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/report/{fileName}": {
            "get": {
                "description": "Download a previously generated CSV report",
                "produces": [
                    "text/csv"
                ],
                "summary": "Download CSV report",
                "operationId": "downloadCsvReport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CSV file name to download",
                        "name": "fileName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file for download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid file name",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/segment/addSegment": {
            "post": {
                "description": "Link user to segments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add user ti a segment",
                "operationId": "addUserToSegment",
                "parameters": [
                    {
                        "description": "Segment object to be added",
                        "name": "segment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.UserSegmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully linked segment to a user",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.UserSegmentRequest"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/segment/new": {
            "post": {
                "description": "Add a new segment to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new segment",
                "operationId": "addSegment",
                "parameters": [
                    {
                        "description": "Segment object to be added",
                        "name": "segment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.SegmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully added segment",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.SegmentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/segment/remove": {
            "delete": {
                "description": "Cascade delete a segment and remove associated users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Cascade delete a segment",
                "operationId": "cascadeDeleteSegment",
                "parameters": [
                    {
                        "description": "Segment object to delete",
                        "name": "segment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.SegmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted segment",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.SegmentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/segment/users/{segmentName}": {
            "get": {
                "description": "Get a list of users belonging to a specific segment",
                "produces": [
                    "application/json"
                ],
                "summary": "Get users of a segment",
                "operationId": "getSegmentUsersInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "segment Name to get list of its users",
                        "name": "segmentName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved segment users",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.GetUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/user/new": {
            "post": {
                "description": "Add a new user to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new user",
                "operationId": "addUser",
                "parameters": [
                    {
                        "description": "User object to be added",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully added user",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/user/segments": {
            "delete": {
                "description": "Remove a user from one or more segments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Remove a user from one or more segments",
                "operationId": "deleteUserFromSegment",
                "parameters": [
                    {
                        "description": "User and segment association",
                        "name": "userSegments",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.UserSegmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully removed user from segment",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.UserSegmentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        },
        "/user/segments/{userID}": {
            "get": {
                "description": "Get information about the segments a user belongs to",
                "produces": [
                    "application/json"
                ],
                "summary": "Get user's segments information",
                "operationId": "getUserSegmentsInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user ID to get list of segments for",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user segments",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.GetSegmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    },
                    "409": {
                        "description": "Query execution failure",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_api.ResponseStatus"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_controller_api.CsvReportRequest": {
            "type": "object",
            "required": [
                "month",
                "year"
            ],
            "properties": {
                "month": {
                    "$ref": "#/definitions/time.Month"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "internal_controller_api.CsvReportResponse": {
            "type": "object",
            "required": [
                "csv_url"
            ],
            "properties": {
                "csv_url": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_controller_api.GetUsersResponse": {
            "type": "object",
            "required": [
                "user_ids",
                "user_segment"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_segment": {
                    "type": "string"
                }
            }
        },
        "internal_controller_api.ResponseStatus": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_controller_api.SegmentRequest": {
            "type": "object",
            "required": [
                "slug"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "internal_controller_api.SegmentResponse": {
            "type": "object",
            "required": [
                "slug"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_controller_api.UserRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_controller_api.UserResponse": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_controller_api.UserSegmentRequest": {
            "type": "object",
            "required": [
                "segment_slug",
                "user_id"
            ],
            "properties": {
                "segment_slug": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_controller_api.UserSegmentResponse": {
            "type": "object",
            "required": [
                "segment_slug",
                "user_id"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "segment_slug": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}